增加Builder来创建加密解密器
