+++ src/main/java/io/xjar/dir/XDirEncryptor.java
@@ -25,6 +25,11 @@ public class XDirEncryptor extends XEntryEncryptor<File> implements XEncryptor {
     }
 
     @Override
+    public void encrypt(XKey key, String src, String dest) throws IOException {
+        encrypt(key, new File(src), new File(dest));
+    }
+
+    @Override
     public void encrypt(XKey key, File src, File dest) throws IOException {
         if (src.isFile()) {
             XEncryptor encryptor = filtrate(src) ? xEncryptor : xNopEncryptor;
